# src/api-contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DailyNagger.Server | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5010/"
    }
  ],
  "paths": {
...
    "/api/task-logs/{id}": {
      "put": {
        "tags": [
          "TaskLogs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonElement"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskLogDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/task-logs/{id}/task-entries": {
      "patch": {
        "tags": [
          "TaskLogs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskEntryValuesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskLogVersionDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ClientIdentityDto": {
        "required": [
          "clientId",
          "deviceName",
          "deviceModel"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          },
          "deviceModel": {
            "type": "string"
          }
        }
      },
      "JsonElement": { },
...
      "TaskEntryDto": {
        "required": [
          "id",
          "taskLogId",
          "parentTaskItemId",
          "label",
          "description",
          "valueType",
          "tag",
          "value",
          "lastTaskRunReferenceValue",
          "rolloverBehavior",
          "interactionAt",
          "interactionTimeZone",
          "interactionLocale",
          "interactionMood",
          "interactionMoodAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "taskLogId": {
            "type": "string",
            "format": "uuid"
          },
          "parentTaskItemId": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "valueType": {
            "$ref": "#/components/schemas/TaskEntryValueTypeDto"
          },
          "tag": {
            "type": [
              "null",
              "string"
            ]
          },
          "value": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastTaskRunReferenceValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "rolloverBehavior": {
            "$ref": "#/components/schemas/TaskEntryRolloverBehaviorDto"
          },
          "interactionAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "interactionTimeZone": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionLocale": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionMood": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionMoodAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "TaskEntryRolloverBehaviorDto": {
        "enum": [
          "Remove",
          "MoveValueToHistory",
          "CarryOverValue"
        ]
      },
      "TaskEntryValueTypeDto": {
        "enum": [
          "Text",
          "Integer",
          "Decimal",
          "Boolean"
        ]
      },
      "TaskEntryValueUpdateDto": {
        "required": [
          "id",
          "value",
          "interactionAt",
          "interactionTimeZone",
          "interactionLocale",
          "interactionMood",
          "interactionMoodAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "interactionTimeZone": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionLocale": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionMood": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionMoodAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "TaskItemDto": {
        "required": [
          "id",
          "taskLogId",
          "parentTaskItemId",
          "name",
          "taskEntries",
          "taskItems",
          "tag",
          "isDone",
          "rolloverBehavior",
          "interactionAt",
          "interactionTimeZone",
          "interactionLocale",
          "interactionMood",
          "interactionMoodAt",
          "descendantTaskItemCount",
          "doneDescendantTaskItemCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "taskLogId": {
            "type": "string",
            "format": "uuid"
          },
          "parentTaskItemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "taskEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskEntryDto"
            }
          },
          "taskItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskItemDto"
            }
          },
          "tag": {
            "type": [
              "null",
              "string"
            ]
          },
          "isDone": {
            "type": "boolean"
          },
          "rolloverBehavior": {
            "$ref": "#/components/schemas/TaskItemRolloverBehaviorDto"
          },
          "interactionAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "interactionTimeZone": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionLocale": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionMood": {
            "type": [
              "null",
              "string"
            ]
          },
          "interactionMoodAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "descendantTaskItemCount": {
            "type": "integer",
            "format": "int32"
          },
          "doneDescendantTaskItemCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TaskItemRolloverBehaviorDto": {
        "enum": [
          "Keep",
          "RemoveWhenDone"
        ]
      },
      "TaskLogDto": {
        "required": [
          "id",
          "nagId",
          "copiedFromTaskLogId",
          "closedOn",
          "updatedAt",
          "updatedByClientId",
          "updatedByDeviceName",
          "updatedByDeviceModel",
          "version",
          "taskItems",
          "tag",
          "descendantTaskItemCount",
          "doneDescendantTaskItemCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nagId": {
            "type": "string",
            "format": "uuid"
          },
          "copiedFromTaskLogId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "closedOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedByClientId": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedByDeviceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedByDeviceModel": {
            "type": [
              "null",
              "string"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "taskItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskItemDto"
            }
          },
          "tag": {
            "type": [
              "null",
              "string"
            ]
          },
          "descendantTaskItemCount": {
            "type": "integer",
            "format": "int32"
          },
          "doneDescendantTaskItemCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TaskLogVersionDto": {
        "required": [
          "version",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateTaskEntryValuesRequest": {
        "required": [
          "communityId",
          "userId",
          "payload",
          "updatedAt",
          "baseVersion",
          "nextVersion"
        ],
        "type": "object",
        "properties": {
          "communityId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "payload": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskEntryValueUpdateDto"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "baseVersion": {
            "type": "integer",
            "format": "int32"
          },
          "nextVersion": {
            "type": "integer",
            "format": "int32"
          },
          "clientIdentity": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ClientIdentityDto"
              }
            ]
          }
        }
      }
    }
  },
...
}